third_party/blink/web_tests/http/tests/devtools/console/console-search-expected.txt

Tests console search.


Running: waitForMessages
Message count: 203

Running: assertViewportHeight

Running: testSearchOutsideOfViewport
Search matches: '200 matches'

Running: scrollConsoleToTop
first visible message index: 0

Running: testCanJumpForward
first visible message index: 0
first visible message index: 0
last visible message index: 202

Running: testCanJumpBackward
last visible message index: 202
first visible message index: 0
first visible message index: 0

Running: scrollConsoleToTop
first visible message index: 0

Running: testCanMarkCurrentMatch
number of visible matches: 2
match 0: highlighted-search-result current-search-result
match 1: highlighted-search-result
number of visible matches: 2
match 0: highlighted-search-result
match 1: highlighted-search-result current-search-result

Running: testCanJumpForwardBetweenTreeElementMatches
matched tree element: field_1
matched tree element: field_199

Running: testCaseInsensitiveRegex
number of visible matches: 2
  match 0: highlighted-search-result current-search-result
  match 1: highlighted-search-result


Running: testCaseSensitiveTextWithoutMatches
number of visible matches: 0


Running: testCaseSensitiveTextWithMatches
number of visible matches: 2
  match 0: highlighted-search-result current-search-result
  match 1: highlighted-search-result


Running: testCaseSensitiveRegexWithoutMatches
number of visible matches: 0


Running: testCaseSensitiveRegexWithMatches
number of visible matches: 2
  match 0: highlighted-search-result current-search-result
  match 1: highlighted-search-result
